RL/requirements.txt
# PyTorch and Deep Learning
torch>=2.0.0
torch-geometric>=2.3.0
numpy>=1.24.0

# Graph Processing
networkx>=3.0

# Kafka Integration
kafka-python>=2.0.2
confluent-kafka>=2.3.0

# Visualization and Plotting
matplotlib>=3.7.0

# Data Processing and Utilities
scipy>=1.10.0

# Optional: Jupyter support (if using notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Optional: Additional ML utilities
scikit-learn>=1.3.0
pandas>=2.0.0

# Optional: Progress bars and logging
tqdm>=4.65.0
